added additional checks to document controllers to handle if document is not found

+++ app/controllers/document_controller.py
@@ -11,14 +11,18 @@ def get_all_documents():
         user = db.session.execute(db.select(User).filter_by(id=current_user)).scalar()
 
         documents = db.session.execute(db.select(Document).filter_by(user_id=current_user)).scalars().all()
-        documents_data = [{
+
+        if documents:
+            documents_data = [{
             'id': document.id,
             'created_at': document.created_at,
             'name': document.name,
             'content': document.content
-        } for document in documents]
+            } for document in documents]
         
-        return jsonify({"documents": documents_data, "user": user.email}), 200
+            return jsonify({"documents": documents_data, "user": user.email}), 200
+        else: 
+            return jsonify({'error': 'Documents not found'}, 404)
     except Exception as e:
         return jsonify({'error': str(e)}), 500
 
@@ -36,7 +40,8 @@ def get_document(document_id):
             'content': document.content
             }
             return jsonify({'document': document_data}), 200
-        
+        else:
+            return jsonify({'error': 'Document not found'}, 404)
     except Exception as e:
         return jsonify({'error': str(e)}), 500
     
@@ -62,11 +67,14 @@ def update_document():
         data = request.get_json()
         document_data = data['updatedDocumentData']
         document = db.session.execute(db.select(Document).filter_by(user_id=current_user, id=document_data["id"])).scalar()
-        document.name = document_data["name"]
-        document.content = document_data["content"]
-        db.session.commit()
 
-        return jsonify('Document has been updated.'), 200
+        if document:
+            document.name = document_data["name"]
+            document.content = document_data["content"]
+            db.session.commit()
+            return jsonify('Document has been updated.'), 200
+        else: 
+            return jsonify({'error': 'Document not found'}, 404)
     except Exception as e:
         db.session.rollback()
         return jsonify({'error': str(e)}), 500
@@ -76,10 +84,13 @@ def delete_document(document_id):
     try:
         current_user = get_jwt_identity()
         document = db.session.execute(db.select(Document).filter_by(user_id=current_user, id=document_id)).scalar()
-        db.session.delete(document)
-        db.session.commit()
 
-        return jsonify('Document has been deleted.'), 204
+        if document:
+            db.session.delete(document)
+            db.session.commit()
+            return jsonify('Document has been deleted.'), 204
+        else: 
+            return jsonify({'error': 'Document not found'}, 404)
     except Exception as e:
         db.session.rollback()
         return jsonify({'error': str(e)}), 500
